Extract FilterPickerBody
