+++ frontend/src/metabase/common/components/FilterPicker/FilterPicker.tsx
@@ -12,15 +12,9 @@ import { isExpression as isLegacyExpression } from "metabase-lib/expressions";
 import LegacyFilter from "metabase-lib/queries/structured/Filter";
 import type LegacyQuery from "metabase-lib/queries/StructuredQuery";
 
-import type { ColumnListItem, SegmentListItem } from "./types";
-
-import { BooleanFilterPicker } from "./BooleanFilterPicker";
-import { DateFilterPicker } from "./DateFilterPicker";
-import { NumberFilterPicker } from "./NumberFilterPicker";
-import { CoordinateFilterPicker } from "./CoordinateFilterPicker";
-import { StringFilterPicker } from "./StringFilterPicker";
-import { TimeFilterPicker } from "./TimeFilterPicker";
 import { FilterColumnPicker } from "./FilterColumnPicker";
+import { FilterPickerBody } from "./FilterPickerBody";
+import type { ColumnListItem, SegmentListItem } from "./types";
 
 export interface FilterPickerProps {
   query: Lib.Query;
@@ -124,25 +118,17 @@ export function FilterPicker({
     );
   }
 
-  const FilterWidget = getFilterWidget(column);
-
-  if (FilterWidget) {
-    return (
-      <FilterWidget
-        query={query}
-        stageIndex={stageIndex}
-        column={column}
-        filter={filter}
-        isNew={isNewFilter}
-        onChange={handleChange}
-        onBack={() => setColumn(undefined)}
-      />
-    );
-  }
-
-  // This codepath should never be hit,
-  // but is here to make TypeScript happy
-  return renderExpressionEditor();
+  return (
+    <FilterPickerBody
+      query={query}
+      stageIndex={stageIndex}
+      column={column}
+      filter={filter}
+      isNew={isNewFilter}
+      onChange={handleChange}
+      onBack={() => setColumn(undefined)}
+    />
+  );
 }
 
 function getInitialColumn(
@@ -165,26 +151,4 @@ function isExpressionEditorInitiallyOpen(
     !Lib.isStandardFilter(query, stageIndex, filter) &&
     !Lib.isSegmentFilter(query, stageIndex, filter)
   );
-}
-
-function getFilterWidget(column: Lib.ColumnMetadata) {
-  if (Lib.isBoolean(column)) {
-    return BooleanFilterPicker;
-  }
-  if (Lib.isTime(column)) {
-    return TimeFilterPicker;
-  }
-  if (Lib.isDate(column)) {
-    return DateFilterPicker;
-  }
-  if (Lib.isCoordinate(column)) {
-    return CoordinateFilterPicker;
-  }
-  if (Lib.isString(column)) {
-    return StringFilterPicker;
-  }
-  if (Lib.isNumeric(column)) {
-    return NumberFilterPicker;
-  }
-  return null;
 }
