+++ frontend/src/metabase/common/components/FilterPicker/FilterPickerBody/index.ts
@@ -1,0 +1,1 @@
+export * from "./FilterPickerBody";
