+++ frontend/src/metabase/common/components/FilterPicker/index.ts
@@ -1,2 +1,3 @@
 export { FilterPicker } from "./FilterPicker";
+export { FilterPickerBody } from "./FilterPickerBody";
 export { SimpleDateFilterPicker } from "./DateFilterPicker";
